requirements.txt
django==3.0.6
plaid_python==8.8.0